AzureFunctions/requirements.txt
azure-functions==1.5.0
ecdsa==0.14.1
pyasn1==0.4.8
python-jose==3.2.0
rsa==4.6
six==1.15.0